make install will not remove but just copy the software bundle, this keeps previously installed depenedencies from packages alive
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@8049 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ build/netbeans/karabo/.package.sh
@@ -32,7 +32,9 @@ PACKAGEDIR=$(pwd)/../../../package/$CONF/$DISTRO_ID/$DISTRO_RELEASE/$MACHINE/$PA
 #INSTALLSCRIPT=karabo-${VERSION}-${CONF}-${DISTRO_ID}-${DISTRO_RELEASE}-${MACHINE}.sh
 INSTALLSCRIPT=${PACKAGENAME}-${CONF}-${DISTRO_ID}-${DISTRO_RELEASE}-${MACHINE}.sh
 
-if [ -d $(pwd)/../../../package/$CONF/$DISTRO_ID/$DISTRO_RELEASE/$MACHINE ]; then rm -rf $(pwd)/../../../package/$CONF/$DISTRO_ID/$DISTRO_RELEASE/$MACHINE; fi
+if [ "$PACKAGE_TYPE" = "tar" ]; then
+    if [ -d $(pwd)/../../../package/$CONF/$DISTRO_ID/$DISTRO_RELEASE/$MACHINE ]; then rm -rf $(pwd)/../../../package/$CONF/$DISTRO_ID/$DISTRO_RELEASE/$MACHINE; fi
+fi
 mkdir -p $PACKAGEDIR
 
 # karabo
